solve comments
