+++ lib/bali/form_builder/file_fields.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Bali
-  module FormBuilderHelpers
+  class FormBuilder < ActionView::Helpers::FormBuilder
     module FileFields
       def file_field_group(method, options = {})
         options.with_defaults!(
