+++ lib/bali/form_builder/number_fields.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 module Bali
-  module FormBuilderHelpers
-    module TextFields
-      def text_field_group(method, options = {})
+  class FormBuilder < ActionView::Helpers::FormBuilder
+    module NumberFields
+      def number_field_group(method, options = {})
         @template.render Bali::FieldGroupWrapper::Component.new self, method, options do
-          text_field(method, options)
+          number_field(method, options)
         end
       end
     end
