+++ lib/bali/form_builder/rich_text_area_fields.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Bali
-  module FormBuilderHelpers
+  class FormBuilder < ActionView::Helpers::FormBuilder
     module RichTextAreaFields
       def rich_text_area_group(method, options = {})
         @template.render Bali::FieldGroupWrapper::Component.new self, method, options do
