+++ lib/bali/form_builder/select_fields.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Bali
-  module FormBuilderHelpers
+  class FormBuilder < ActionView::Helpers::FormBuilder
     module SelectFields
       def select_group(method, values, options = {}, html_options = {})
         @template.render Bali::FieldGroupWrapper::Component.new self, method, options do
