+++ lib/bali/form_builder/submit_fields.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Bali
-  module FormBuilderHelpers
+  class FormBuilder < ActionView::Helpers::FormBuilder
     module SubmitFields
       def submit(value, options = {})
         options.with_defaults!(
