+++ lib/bali/form_builder/text_fields.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 module Bali
-  module FormBuilderHelpers
-    module NumberFields
-      def number_field_group(method, options = {})
+  class FormBuilder < ActionView::Helpers::FormBuilder
+    module TextFields
+      def text_field_group(method, options = {})
         @template.render Bali::FieldGroupWrapper::Component.new self, method, options do
-          number_field(method, options)
+          text_field(method, options)
         end
       end
     end
